pyproject.toml
[project]
name = "DeepSearchAgents"
version = "0.3.2.dev"
description = "🤗 Awesome CodeAct Deep Web Search Agents framework"
authors = [
    {name = "ArthurWang", email = "deepsearchagent4dev@gpt4coursera.com"},
]
dependencies = [
    "fastapi",
    "fastmcp",
    "sse-starlette>=2.3.3",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.29.0",
    "starlette",
    "pydantic>=2.11.3",
    "pydantic_core>=2.33.1",
    "pydantic-settings>=2.9.1",
    "PyYAML>=6.0",
    "requests",
    "python-dotenv",
    "jsonschema>=4.0.0,<5.0.0",
    "smolagents>=1.19.0,<1.20.0",                           # Adapted for v1.20.0
    "smolagents[e2b]",
    "smolagents[docker]",
    "smolagents[telemetry]",
    "smolagents[vision]",
    "smolagents[audio]",
    "smolagents[litellm]",
    "smolagents[toolkit]",
    "smolagents[gradio]",
    "smolagents[mcp]",
    "litellm>=1.73.6",
    "openai>=1.93.1",
    "langchain<=0.3.26",
    "langchain-text-splitters>=0.3.8",
    "torch",
    "wikipedia-api>=0.8.1",    
    "wolframalpha",
    "arxiv",
    "rich>=13.0.0",
    "loguru",
    "nest_asyncio",
    "click>=8.0",
    "typing_extensions",
    "typing-inspection",
    "annotated_types",
    "anyio",
    "sniffio",
    "rich>=14.0.0",
    "prompt-toolkit>=3.0.0",
    "tomli>=2.2.1; python_version < '3.11'",
    "websockets>=15.0.1",
    "xai-sdk>=1.0.0",
    "exa-py>=1.14.15",                                      # [optional]
    "mistralai",                                            # [optional]
    "firecrawl-py",                                         # [optional] 
    "tabulate2",                                            # CodeAgent-sandbox package
    "PyMuPDF",                                              # CodeAgent-sandbox package
    "markitdown",                                           # CodeAgent-sandbox package
#   "futurehouse-client",  # Temporarily disabled due to dm-tree 0.1.8 build issue with CMake 4.0.2
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
override-dependencies = [
    "dm-tree==0.1.9",  # Override to fix CMake 4.0.2 compatibility issue
    "packaging>=25.0",  # Override to resolve conflict between langchain-text-splitters and xai-sdk
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "build",
    "pyproject-hooks",
    "ruff>=0.12.0",
    "coverage>=7.9.2",
    "iniconfig>=2.1.0",
    "pluggy>=1.6.0",
    "pytest>=8.4.1",
    "pytest-asyncio==1.0.0",
    "pytest-cov==6.2.1",
    "pytest-timeout==2.4.0",
]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
cli = [
    "rich>=14.0.0",
    "prompt-toolkit>=3.0.0", 
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
